Upstream version 1.25.33

+++ tests/disks/test-qemu-drive.sh
@@ -47,12 +47,20 @@ rm -f "$DEBUG_QEMU_FILE"
 # Ceph (RBD).
 
 $guestfish <<EOF ||:
-  add "/abc-def/ghi-jkl" "format:raw" "protocol:rbd" \
+  add "abc-def/ghi-jkl" "format:raw" "protocol:rbd" \
     "server:1.2.3.4:1234 1.2.3.5:1235 1.2.3.6:1236"
   run
 EOF
 check_output
 grep -sq -- '-drive file=rbd:abc-def/ghi-jkl:mon_host=1.2.3.4\\:1234\\;1.2.3.5\\:1235\\;1.2.3.6\\:1236:auth_supported=none,' "$DEBUG_QEMU_FILE" || fail
+rm "$DEBUG_QEMU_FILE"
+
+$guestfish <<EOF ||:
+  add "abc-def/ghi-jkl" "format:raw" "protocol:rbd"
+  run
+EOF
+check_output
+grep -sq -- '-drive file=rbd:abc-def/ghi-jkl:auth_supported=none,' "$DEBUG_QEMU_FILE" || fail
 rm "$DEBUG_QEMU_FILE"
 
 # HTTP.
@@ -68,7 +76,7 @@ rm "$DEBUG_QEMU_FILE"
 # Gluster.
 
 $guestfish <<EOF ||:
-  add "/volname/image" "format:raw" "protocol:gluster" "server:www.example.com:24007"
+  add "volname/image" "format:raw" "protocol:gluster" "server:www.example.com:24007"
   run
 EOF
 check_output
@@ -78,7 +86,7 @@ rm "$DEBUG_QEMU_FILE"
 # iSCSI.
 
 $guestfish <<EOF ||:
-  add "/target-iqn-name/lun" "format:raw" "protocol:iscsi" "server:www.example.com:3000"
+  add "target-iqn-name/lun" "format:raw" "protocol:iscsi" "server:www.example.com:3000"
   run
 EOF
 check_output
@@ -106,7 +114,7 @@ rm "$DEBUG_QEMU_FILE"
 # Sheepdog.
 
 $guestfish <<EOF ||:
-  add "/volume" "format:raw" "protocol:sheepdog"
+  add "volume" "format:raw" "protocol:sheepdog"
   run
 EOF
 check_output
